Repository add complete

+++ src/main/java/com/example/bdcource/entity/ReportTypesEntity.java
@@ -13,7 +13,7 @@ public class ReportTypesEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "type_id")
-    private int typeId;
+    private short typeId;
     @Column(name = "type", unique = true)
     private String reportType;
     @Column(name = "action")
